feat: 수익률 계산

+++ src/main/java/lotto/domain/Money.java
@@ -3,31 +3,31 @@ package lotto.domain;
 import lotto.exception.MoneyExceptionType;
 
 public class Money {
-    private static final int MINIMUM_MONEY = 1000;
-    private static final int MAXIMUM_MONEY = 100000;
-    private static final int LOTTO_PRIZE = 1000;
+    private static final long MINIMUM_MONEY = 1000;
+    private static final long MAXIMUM_MONEY = 100000;
+    private static final long LOTTO_PRIZE = 1000;
 
-    private final int money;
+    private final long money;
 
-    public Money(int money) {
+    public Money(long money) {
         validateRange(money);
         validateDivision(money);
         this.money = money;
     }
 
-    private void validateRange(int money) throws IllegalArgumentException {
+    private void validateRange(long money) throws IllegalArgumentException {
         if (money < MINIMUM_MONEY || money > MAXIMUM_MONEY) {
             throw new IllegalArgumentException(MoneyExceptionType.OUT_OF_RANGE_MONEY.getMessage());
         }
     }
 
-    private void validateDivision(int money) throws IllegalArgumentException {
+    private void validateDivision(long money) throws IllegalArgumentException {
         if (money % LOTTO_PRIZE != 0) {
             throw new IllegalArgumentException(MoneyExceptionType.NON_DIVISIBLE_MONEY.getMessage());
         }
     }
 
-    public int getMoney() {
+    public long getMoney() {
         return money;
     }
 }
